% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ages.R
\name{bin_ages}
\alias{bin_ages}
\title{Break time data up into bins}
\usage{
bin_ages(x, by = NULL, number = NULL, age = FALSE)
}
\arguments{
\item{x}{vector of ages}

\item{by}{bin width}

\item{age}{logical bin age returned, not number (default FALSE, return bin number)}
}
\value{
vector of bin memberships
}
\description{
Have fun with this. Basic rules. Greater than equal to base, less than top.
}
\author{
Peter D Smits \href{mailto:peterdavidsmits@gmail.com}{peterdavidsmits@gmail.com}
}
