% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithology_foo.R
\name{simplify_lithology}
\alias{simplify_lithology}
\title{Extract summary of lithological composition (strict)}
\usage{
simplify_lithology(lith)
}
\arguments{
\item{lith}{a vector of lithological descriptions like thos exported by Macrostrat.}
}
\value{
list each element is a unit's simple lithological description
}
\description{
Given a textual description of a geological unit's lithology, extract the key terms.
}
