% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_dup.R
\name{worddup}
\alias{worddup}
\title{Remove duplicated words from a vector of strings}
\usage{
worddup(desc)
}
\arguments{
\item{desc}{nested list of descriptions}
}
\value{
desc but duplicated words removed
}
\description{
Given a (nested) list of descriptions, remove the duplicated words.
}
