% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_rep.R
\name{wordrep}
\alias{wordrep}
\title{Synonmize terms}
\usage{
wordrep(desc, wordd)
}
\arguments{
\item{desc}{description}

\item{wordd}{vector of synonymous words}
}
\value{
description vector with synonyms replaced (no changes if no synonyms.)
}
\description{
Replace words in lithological description with one word
}
