% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_rm.R
\name{wordrm}
\alias{wordrm}
\title{Remove words from lithological description}
\usage{
wordrm(desc, words)
}
\arguments{
\item{desc}{string description}

\item{words}{vector of strings to be removed from desc}
}
\value{
desc but with words removed
}
\description{
Given a string, remove occurrences of certain words.
}
